pyproject.toml
[build-system]
requires = ["uv_build>=0.9.6,<0.10.0"]
build-backend = "uv_build"

[project]
name = "lfp-build"
version = "0.0.1+g19a292c"
description = "Local build utilities"
requires-python = ">=3.11,<3.14"
dependencies = [
    "basedpyright",
    "uv>=0.9.9",
    "libcst",
    "mergedeep",
    "ruff",
    "tomlkit",
    "cyclopts",
    "lfp-logging @ git+https://github.com/regbo/lfp-logging-py.git",
    "python-dotenv",
    "sitecustomize-entrypoints",
]

[project.scripts]
lfp-build = "lfp_build.cli:main"

[project.entry-points.sitecustomize]
lfp-build-config = "lfp_build._config:load"

# test123
[dependency-groups]
dev = ["pytest"]


[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "ANN"]
ignore = ["ANN401"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["ANN001", "ANN002", "ANN003", "ANN202", "ANN201"]